+++ R/RaceModels.R
@@ -104,7 +104,7 @@
 #' set to 0.
 #'
 #' @note Similarly to the drift diffusion models (like \code{ddiffusion} and
-#' \code{\link{dWEV}}), `s1` and `s2` are scaling factors (`s1` scales: \code{mu1} and  \code{a},
+#' \code{\link{ddynaViTE}}), `s1` and `s2` are scaling factors (`s1` scales: \code{mu1} and  \code{a},
 #' `s2` scales: \code{mu2} and \code{b}, and depending on response: if `response=2`, `s1` scales
 #' \code{th1},\code{th2},and \code{wrt}), otherwise `s2` is the scaling factor. It is sometimes
 #' assumed (Moreno-Bote, 2010), that both noise terms are equal, then they should definitely be