+++ R/fitRTConf.R
@@ -5,7 +5,7 @@
 #' flavors of race models (Hellmann et al., 2023). Which model to fit is
 #' specified by the argument \code{model}.
 #' Only a ML method is implemented.
-#' See \code{\link{dWEV}}, \code{\link{d2DSD}}, and \code{\link{dRM}} for more
+#' See \code{\link{ddynaViTE}}, \code{\link{d2DSD}}, and \code{\link{dRM}} for more
 #' information about the parameters and Details for not-fitted parameters.
 #'
 #' @param data a `data.frame` where each row is one trial, containing following
@@ -55,9 +55,9 @@
 #' (default) the number of available cores -1 is used.
 #' @param restr_tau numerical or `Inf` or `"simult_conf"`. For 2DSD and dynWEV only.
 #' Upper bound for tau. Fits will be in the interval (0,`restr_tau`). If FALSE tau will be unbound.
-#' For `"simult_conf"`, see the documentation of \code{\link{d2DSD}} and \code{\link{dWEV}}
+#' For `"simult_conf"`, see the documentation of \code{\link{d2DSD}} and \code{\link{ddynaViTE}}
 #' @param precision numeric. Precision of calculation.
-#' (in the respective models) for the density functions (see \code{\link{dWEV}} for more information).
+#' (in the respective models) for the density functions (see \code{\link{ddynaViTE}} for more information).
 #' @param ... Possibility of giving alternative variable names in data frame
 #' (in the form \code{condition = "SOA"}, or \code{response="pressedKey"}).
 #'
@@ -157,10 +157,10 @@
 #' # We use one of the implemented models, "dynWEV"
 #' # 1. Generate data
 #' # data with positive drift (stimulus = "upper")
-#' data <- rWEV(20, a=2,v=0.5,t0=0.2,z=0.5, sz=0.1,sv=0.1, st0=0,  tau=4, s=1, w=0.3)
+#' data <- rdynaViTE(20, a=2,v=0.5,t0=0.2,z=0.5, sz=0.1,sv=0.1, st0=0,  tau=4, s=1, w=0.3)
 #' data$stimulus <- "upper"
 #' # data with negtive drift (stimulus = "lower") but same intensity
-#' data2 <- rWEV(100, a=2,v=-0.5,t0=0.2,z=0.5,sz=0.1,sv=0.1, st0=0,  tau=4, s=1, w=0.3)
+#' data2 <- rdynaViTE(100, a=2,v=-0.5,t0=0.2,z=0.5,sz=0.1,sv=0.1, st0=0,  tau=4, s=1, w=0.3)
 #' data2$stimulus <- "lower"
 #' data <- rbind(data, data2)
 #' # Transfer response column and add dummy condition column